Calculation and Log implementations

+++ FriendCoordinateCalculatorAPI/Application/ViewModels/CoordinateViewModel.cs
@@ -2,12 +2,12 @@ namespace Application.ViewModels
 {
     public class CoordinateViewModel
     {
-        public CoordinateViewModel(decimal _longitude, decimal _latitude)
+        public CoordinateViewModel(double _longitude, double _latitude)
         {
             Longitude = _longitude;
             Latitude = _latitude;
         }
-        public decimal Longitude { get; set; }
-        public decimal Latitude { get; set; }
+        public double Longitude { get; set; }
+        public double Latitude { get; set; }
     }
 }
